Fix check_whitespace.c to end Memory Leak Problem
check_whitespace has been fixed by adding an if statement to is_cleaned before the return statement. Essentially, the if statement checks to see if the string cleaned is empty or not. If it is empty, nothing happens. But, if it is not empty, the space allocated with result shall be cleared with the method free(). Since the strings are not manipulated after is_cleaned is ran, it is fair to say that the change is safe.